fix: fix version showed in ethtool
Fix the version in the output of ethtool commands from empty to the
kernel version. Use the macro HINIC3_NIC_DRV_VERSION to record driver
changes for developers. The last number 1.0.X grow as minor changes,
like this revision which does not affect the main functions. The
middle number 1.X.1 iterates with changes to drivers functionality
and fixing of serious bugs, and the first one X.0.1 means the number
of releases of the driver.

Signed-off-by: Shi Jing <shijing34@huawei.com>