xfs: forcealign not compatible with reflink and realtime device
hulk inclusion
category: feature
bugzilla: https://gitee.com/openeuler/kernel/issues/I9VTE3
CVE: NA

--------------------------------

Reflink will not be supported for forcealign yet, so disallow a mount
under this condition. This is because we have the limitation of pageache
writeback not knowing how to writeback an entire allocation unut, after
covert extent form cowfork to datafork, force alignment constraints may
be break, so reject a mount with relink.

RT vol will not be supported for forcealign yet, so disallow a mount under
this condition. It will be possible to support RT vol and forcealign in
future. For this, the inode extsize must be a multiple of rtextsize - this
is enforced already in xfs_ioctl_setattr_check_extsize() and
xfs_inode_validate_extsize().

Signed-off-by: Long Li <leo.lilong@huawei.com>