arm64: Faster SVC exception handler with xcall
hulk inclusion
category: feature
bugzilla: https://gitee.com/openeuler/release-management/issues/IB6JLE

--------------------------------

This patch is designed to optimize the performance of the SVC exception
handler by simplifying its operation, which can lead to faster execution
times. However, this optimization comes with the trade-off of reduced
functionality, particularly in areas related to security and maintenance.

When a task is executed with xcall, certain features that are crucial
for robust system operation may not be available, which could impact the
system's ability to perform essential tasks.

Here's a breakdown of the potential impacts:

1. Memory Tagging Extension (MTE)
2. Process Trace (PTRACE)
3. System Call Trace (STRACE)
4. GNU Debugger (GDB)
5. Software single-stepping
6. Secure State Buffer Descriptor (SSBD)
7. Shadow Call Stack
8. Software Translation Table Buffer Zero Protection (SW_TTBR0_PAN)
9. Unmap Kernel at Exception Level 0 (UNMAP_KERNEL_AT_EL0)
10.ARM64_WORKAROUND_SPECULATIVE_UNPRIV_LOAD
11. GCC Plugin Stack Leak Detection (GCC_PLUGIN_STACKLEAK)
12. SYSCALL Trace Point

In conclusion, while the patch is intended to enhance the performance of
the SVC exception handler, it does so by sacrificing several important
features that contribute to security, debugging, and overall system
stability. It is imperative for developers and system administrators to
be cognizant of these trade-offs and to plan for the potential effects
on their applications and operational workflows.

Signed-off-by: Yipeng Zou <zouyipeng@huawei.com>