nfs: Ensure write and flush consume writeback errors
HULK inclusion
category: bugfix
bugzilla: https://gitee.com/openeuler/kernel/issues/IB7MNY

--------------------------------

When the file is closed and the ->flush is executed, the writeback
error is not consumed. As a result, when the file is written next time,
the writeback error still exists, causing a new write failure.

Another problem is that in the write process, only special writeback
error codes such as -EDQUOT, -EFBIG, and -ENOSPC are detected, and the
error codes are consumed. In this case, other error codes may be left.

The writeback error code should be consumed unconditionally in the
write(), fsync(), and close() to avoid sampling residual expired
codes that are left over from the last invoking.

In the fsync process, the error code can be consumed. Therefore, this
patch ensures that write and flush can consume error codes.

Fixes: 6fbda89b257f ("NFS: Replace custom error reporting mechanism with generic one")
Signed-off-by: Wang Zhaolong <wangzhaolong1@huawei.com>
Signed-off-by: Yang Erkun <yangerkun@huawei.com>